feedback-llm-service/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
requests
google-generativeai

